pyproject.toml
[project]
name = "modern-software-dev-assignments"
version = "0.1.0"
description = "Course assignments with FastAPI + SQLite starter and tooling"
authors = [{ name = "Course", email = "course@example.com" }]
requires-python = ">=3.10,<4.0"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "ollama>=0.5.3,<0.6",
    "mcp>=1.1.0",
    "httpx>=0.27.0",
    "hypercorn>=0.16.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "black>=24.1.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
]

[tool.uv]
package = false
default-groups = "all"

[tool.uv.build-backend]
module-root = ""

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]
